double max_bytes_for_level_multiplier_additional

+++ include/rocksdb/advanced_options.h
@@ -664,8 +664,8 @@ struct AdvancedColumnFamilyOptions {
   // Default: 1
   //
   // Dynamically changeable through SetOptions() API
-  std::vector<int> max_bytes_for_level_multiplier_additional =
-      std::vector<int>(num_levels, 1);
+  std::vector<double> max_bytes_for_level_multiplier_additional =
+      std::vector<double>(num_levels, 1);
 
   // We try to limit number of bytes in one compaction to be lower than this
   // threshold. But it's not guaranteed.
